// apps/frontend/src/neevo/enums/button.enum.ts
/** Enum for Button Levels */
export enum ButtonLevels {
	BLOCK = "block",
	INLINE = "inline",
}

/** Enum for Button Icon Position */
export enum ButtonIconPosition {
	CENTER = "center",
	END = "end",
}

/** Datatype enum for Button sizes */
export enum ButtonSizes {
	SMALL = "small",
	MEDIUM = "medium",
	LARGE = "large",
	XLARGE = "xlarge",
	XXLARGE = "xxlarge",
}

/** Enum for Button types */
export enum ButtonTypes {
	BUTTON = "button",
	SUBMIT = "submit",
}